[Refactor] #4755 DungeonRecords::build_known_dungeons() をDungeonService へ移した
