+++ src/system/dungeon/dungeon-record.cpp
@@ -7,28 +7,10 @@
 #include "system/dungeon/dungeon-record.h"
 #include "locale/language-switcher.h"
 #include "system/angband-exceptions.h"
-#include "system/dungeon/dungeon-definition.h"
-#include "system/dungeon/dungeon-list.h"
 #include "system/enums/dungeon/dungeon-id.h"
 #include "term/z-rand.h"
 #include "util/enum-converter.h"
 #include "util/enum-range.h"
-
-namespace {
-std::string get_dungeon_format(DungeonMessageFormat dmf)
-{
-    switch (dmf) {
-    case DungeonMessageFormat::DUMP:
-        return _("   %c%-12s: %3d 階\n", "   %c%-16s: level %3d\n");
-    case DungeonMessageFormat::KNOWLEDGE:
-        return _("%c%-12s :  %3d 階\n", "%c%-16s :  level %3d\n");
-    case DungeonMessageFormat::RECALL:
-        return _("      %c) %c%-12s : 最大 %d 階", "      %c) %c%-16s : Max level %d");
-    default:
-        THROW_EXCEPTION(std::logic_error, format("Invalid dungeon message format is specified! %d", enum2i(dmf)));
-    }
-}
-}
 
 bool DungeonRecord::has_entered() const
 {
@@ -149,41 +131,6 @@ void DungeonRecords::reset_all()
     for (auto &[_, record] : this->records) {
         record->reset();
     }
-}
-
-std::vector<std::string> DungeonRecords::build_known_dungeons(DungeonMessageFormat dmf) const
-{
-    const auto fmt = get_dungeon_format(dmf);
-    const auto &dungeons = DungeonList::get_instance();
-    std::vector<std::string> recall_dungeons;
-    auto num_entered = 0;
-    for (const auto &[dungeon_id, record] : this->records) {
-        if (!record->has_entered()) {
-            continue;
-        }
-
-        const auto max_level = record->get_max_level();
-        const auto &dungeon = dungeons.get_dungeon(dungeon_id);
-        const auto is_dungeon_conquered = dungeon.is_conquered();
-        const auto is_conquered = is_dungeon_conquered || (max_level == dungeon.maxdepth);
-        std::string message;
-        switch (dmf) {
-        case DungeonMessageFormat::DUMP:
-        case DungeonMessageFormat::KNOWLEDGE:
-            message = format(fmt.data(), is_conquered ? '!' : ' ', dungeon.name.data(), max_level);
-            break;
-        case DungeonMessageFormat::RECALL:
-            message = format(fmt.data(), static_cast<char>('a' + num_entered), is_conquered ? '!' : ' ', dungeon.name.data(), max_level);
-            num_entered++;
-            break;
-        default:
-            THROW_EXCEPTION(std::logic_error, format("Invalid dungeon message format is specified! %d", enum2i(dmf)));
-        }
-
-        recall_dungeons.push_back(message);
-    }
-
-    return recall_dungeons;
 }
 
 std::vector<DungeonId> DungeonRecords::collect_entered_dungeon_ids() const
