+++ src/system/services/dungeon-service.h
@@ -10,6 +10,13 @@
 #include <memory>
 #include <optional>
 #include <string>
+#include <vector>
+
+enum class DungeonMessageFormat {
+    DUMP,
+    KNOWLEDGE,
+    RECALL,
+};
 
 enum class DungeonId;
 class DungeonDefinition;
@@ -20,4 +27,5 @@ public:
     static int decide_gradiator_level();
     static int find_max_level();
     static std::optional<std::string> check_first_entrance(DungeonId dungeon_id);
+    static std::vector<std::string> build_known_dungeons(DungeonMessageFormat dmf);
 };
